Add more attributes to user and more users

+++ user-service/src/main/java/de/tine/cdc/demo/userservice/UserController.java
@@ -14,7 +14,7 @@ public class UserController {
     }
 
     @GetMapping("/users/{userId}")
-    public User getUser(@PathVariable Long userId) {
+    public User getUser(@PathVariable Integer userId) {
         return userService.findUser(userId);
     }
